Implement the simulator and various constants
The Singleton pattern we have been using until now doesn't seem to work
properly, and therefore the the simulation handler object was put in the
utils.Constants module.

Also, in order to display a second windows with the CV_GUI_Handler,
methods have been added to the CV Gui Handler.